backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
supabase==2.7.4
transformers==4.40.0
torch==2.3.0
httpx==0.27.0
slowapi==0.1.9
python-multipart==0.0.9
pydantic-settings==2.2.1
email-validator==2.1.1
